velocity measurement kind of working

+++ gui/qt_reader_threads/lidar_reader_thread.py
@@ -6,7 +6,7 @@ import queue
 
 class LidarReaderThread(QThread):
     onUpdateLidarDistance = pyqtSignal(int)
-    onUpdateLidarVelocity = pyqtSignal(int)
+    onUpdateLidarVelocity = pyqtSignal(float)
 
     SAVE_FILE_NAME = 'lidar_data.csv'
